fix: Iterator on immutable bitmap didn't work

+++ src/main/java/org/roaringbitmap/buffer/ImmutableRoaringBitmap.java
@@ -1088,7 +1088,7 @@ public class ImmutableRoaringBitmap
 
       ShortIterator iter;
 
-      short pos = 0;
+      int pos = 0;
 
       int x;
